Add option to display image in the alertview.
Add option to display an image in the alertview
for things like a 5-star icon or an app icon.
The image will display centered directly under the
alertview message text. 

Refactor localization to use predefined keys for 
localized lookups rather than using the entire
text string -- this makes the code more readable
and easier to implement changes to the strings.