Check an additional string the bundle for appName.
Add an additional custom key check in the main bundle to determine the app name. The purpose of the custom key check is to allow developers to specify a full name in the rating alert for app names which are too long to include as the app display name (the name that appears under the icon on the device). The name displayed in the alert will be the first non-null result of the following bundle key checks.

1. CFAppStoreDisplayName - This is a custom key that the developer can add to their plist. This key is also useful in other cases where you want to display the apps full name rather than the icon abbreviated name, such as on an info/about screen.

2. CFBundleDisplayName - This is the name that appears under the icon on the device.

3. CFBundleName - The name of the bundle. Typically this is the same as CFBundleDisplayName.