Redesigned to show prompt only when called directly.
Note: This is a breaking change. The UIAlertView will not longer be displayed automatically. After calling +appLaunchedWithID: or +applicationWillEnterForeground: you _must_ call -showPromptIfNeeded wherever you want the prompt to appear -- this would be directly after the aforementioned class methods if you want Appirater to work the same way it did previously.

The purpose of this change is to allow more flexibility as to when the rating prompt is displayed in your app. There are many use cases where you may not want it to display the app rating alert view as soon as the app is launched -- for example if you have a login screen that the user needs to pass first, or if you have other alerts that need to be displayed before the rating prompt. 

-showPromptIfNeeded will show the app rating alert if the app passes all the required validations.