Fix critical bug in showPromptIfNeeded.
showPromptIfNeeded would incorrectly always show the debug version 
of the prompt every time it is called. This occurred because APPIRATER_DEBUG 
was originally only defined if DEBUG was defined. However, it has
since been refactored to be used as a boolean variable. This bit of
code was still using an #ifdef check.